% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output_vaccine.R
\name{add_age}
\alias{add_age}
\title{Keep age groups}
\usage{
add_age(x)
}
\arguments{
\item{x}{age-disaggregated odin output matrix}
}
\value{
age-disaggregated output matrix
}
\description{
Keep age groups
}
