% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{adjusted_eigen}
\alias{adjusted_eigen}
\title{Compute age-adjusted eigenvalue for mixing matrix}
\usage{
adjusted_eigen(dur_IMild, dur_ICase, prob_hosp, mixing_matrix)
}
\arguments{
\item{dur_IMild}{Duration of mild infectiousness (days)}

\item{dur_ICase}{Delay between symptom onset and requiring hospitalisation (days)}

\item{prob_hosp}{Probability of hospitilisation by ages}

\item{mixing_matrix}{Mixing matrix}
}
\value{
Eigenvalue
}
\description{
Compute age-adjusted eigenvalue for mixing matrix
}
