% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{assess_data_quality}
\alias{assess_data_quality}
\title{Assess Data Quality}
\usage{
assess_data_quality(data, treatments, min_timepoints, min_replicates)
}
\arguments{
\item{data}{Experimental data frame}

\item{treatments}{Treatment names vector}

\item{min_timepoints}{Minimum required timepoints}

\item{min_replicates}{Minimum required replicates}
}
\value{
List with data quality assessment
}
\description{
Check data quality requirements for optimization
}
