% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta_est}
\alias{beta_est}
\title{Estimate beta for a squire model using model parameters}
\usage{
beta_est(squire_model, model_params, R0)
}
\arguments{
\item{squire_model}{A squire model. Default = \code{explicit_SEEIR()}}

\item{model_params}{Squire model parameters, from a call to
\code{generate_parameters()}.}

\item{R0}{Basic reproduction number}
}
\value{
Beta parameter
}
\description{
Estimate beta for a squire model using model parameters
}
