% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta_est_env_explicit}
\alias{beta_est_env_explicit}
\title{Estimate beta parameter for explicit environmental model}
\usage{
beta_est_env_explicit(
  dur_IMild,
  dur_ICase,
  prob_hosp,
  mixing_matrix,
  R0,
  env_slp,
  env_dat
)
}
\arguments{
\item{dur_IMild}{Duration of mild infectiousness (days)}

\item{dur_ICase}{Delay between symptom onset and requiring hospitalisation (days)}

\item{prob_hosp}{Probability of hospitilisation by ages}

\item{mixing_matrix}{Mixing matrix}

\item{R0}{Basic reproduction number}

\item{env_slp}{Slope of environmental effect on R0}

\item{env_dat}{Daily environmental data}
}
\value{
Beta parameter
}
\description{
Estimate beta parameter for explicit environmental model
}
