% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta_est_explicit}
\alias{beta_est_explicit}
\title{Estimate beta parameter}
\usage{
beta_est_explicit(dur_IMild, dur_ICase, prob_hosp, mixing_matrix, R0)
}
\arguments{
\item{dur_IMild}{Duration of mild infectiousness (days)}

\item{dur_ICase}{Delay between symptom onset and requiring hospitalisation (days)}

\item{prob_hosp}{Probability of hospitilisation by ages}

\item{mixing_matrix}{Mixing matrix}

\item{R0}{Basic reproduction number}
}
\value{
Beta parameter
}
\description{
Estimate beta parameter
}
