% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta_est_simple}
\alias{beta_est_simple}
\title{Estimate beta parameter}
\usage{
beta_est_simple(duration_infectiousness, mixing_matrix, R0)
}
\arguments{
\item{duration_infectiousness}{Duration of infectiousness (days)}

\item{mixing_matrix}{Mixing matrix}

\item{R0}{Basic reproduction number}
}
\value{
Beta parameter
}
\description{
Estimate beta parameter
}
