% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{calculate_response_metric}
\alias{calculate_response_metric}
\title{Calculate Response Metric}
\usage{
calculate_response_metric(data)
}
\arguments{
\item{data}{Experimental data frame}
}
\value{
Vector of response metrics
}
\description{
Calculate appropriate response metric for statistical testing
}
