% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate Model}
\usage{
calibrate(data, country, replicates = 100, ...)
}
\arguments{
\item{data}{Data frame with 2 variables: date and deaths}

\item{country}{Character. Country data originates from.}

\item{replicates}{Simulation Repetitions. Default = 100}

\item{...}{Other parameters to pass to \code{\link{run_explicit_SEEIR_model}}}
}
\value{
Long data frame of simulation replicates
}
\description{
Calibrate Model
}
\details{
Fit the explicit_SEEIR model to time series of deaths
}
