% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{calibrate_galahad_parameters}
\alias{calibrate_galahad_parameters}
\title{Calibrate GALAHAD Parameters}
\usage{
calibrate_galahad_parameters(
  data,
  treatments,
  response_type,
  validation_results,
  verbose = FALSE
)
}
\arguments{
\item{data}{Experimental data frame}

\item{treatments}{Treatment names vector}

\item{response_type}{Type of biological response}

\item{validation_results}{Results from statistical validation}

\item{verbose}{Print progress messages}
}
\value{
Calibrated GALAHAD configuration
}
\description{
Two-cycle parameter calibration for geometry-adaptive optimization
}
