% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{calibrate_optimization_parameters}
\alias{calibrate_optimization_parameters}
\title{Calibrate Optimization Parameters}
\usage{
calibrate_optimization_parameters(data, geometry_config)
}
\arguments{
\item{data}{Experimental data frame}

\item{geometry_config}{Geometry configuration}
}
\value{
Optimization configuration
}
\description{
Calibrate numerical parameters for optimization
}
