% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_stripped}
\alias{calibrate_stripped}
\title{Calibrate Model}
\usage{
calibrate_stripped(data, country, replicates = 10, ...)
}
\arguments{
\item{data}{Data frame with 2 variables: date and deaths}

\item{country}{Character. Country data originates from.}

\item{replicates}{Simulation Repetitions. Default = 10}

\item{...}{Other parameters to pass to \code{\link{run_explicit_SEEIR_model}}}
}
\value{
List of unformatted odin outputs with the date
}
\description{
Calibrate Model
}
\details{
Fit the explicit_SEEIR model to time series of deaths
}
