% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_time_change}
\alias{check_time_change}
\title{Check time change inputs are correct}
\usage{
check_time_change(tt, time_period)
}
\arguments{
\item{tt}{Time change points}

\item{time_period}{Length of simulation}
}
\value{
Nothing if check pass
}
\description{
Check time change inputs are correct
}
