% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{collapse_age}
\alias{collapse_age}
\title{Collapse age groups in output}
\usage{
collapse_age(d)
}
\arguments{
\item{d}{output data.frame}
}
\value{
Output data.frame
}
\description{
Sums over age groups for each compartment
}
