% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{collapse_compartment}
\alias{collapse_compartment}
\title{Collapse compartments in output}
\usage{
collapse_compartment(d)
}
\arguments{
\item{d}{output data.frame}
}
\value{
Output data.frame
}
\description{
Sums over simplified groups across compartments
}
