% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{collapse_for_report}
\alias{collapse_for_report}
\title{Collapse compartments for reporting major reporters of epidemic}
\usage{
collapse_for_report(d)
}
\arguments{
\item{d}{output data.frame}
}
\value{
Output data.frame
}
\description{
Sums over simplified groups (ICU demand, ICU occupance, hospital demand,
hospital occupancy, infections, deaths)
}
