% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{compile_optimization_results}
\alias{compile_optimization_results}
\title{Compile Optimization Results}
\usage{
compile_optimization_results(treatment_parameters, treatments)
}
\arguments{
\item{treatment_parameters}{List of treatment-specific parameters}

\item{treatments}{Vector of treatment names}
}
\value{
Compiled optimization results
}
\description{
Compile results from multiple treatment optimizations
}
