% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contact_matrices}
\alias{contact_matrices}
\title{Contact matrices}
\format{
A list of contact matrices
}
\usage{
contact_matrices
}
\description{
A list of contact matrices representing different ...
}
\keyword{datasets}
