% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_specific_healthcare_capacity}
\alias{country_specific_healthcare_capacity}
\title{Country specific healthcare capacity}
\format{
A data.frame
}
\usage{
country_specific_healthcare_capacity
}
\description{
Hospital and ICU beds per 1000 pop for various countries
}
\keyword{datasets}
