% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc.R
\name{create_master_chain}
\alias{create_master_chain}
\title{create a master chain from a pmcmc_list object}
\usage{
create_master_chain(x, burn_in)
}
\arguments{
\item{x}{a pmcmc_list object}

\item{burn_in}{an integer denoting the number of samples to discard from each chain}
}
\description{
create a master chain from a pmcmc_list object
}
