% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderly_plotting.R
\name{death_data_format}
\alias{death_data_format}
\title{Generate death time seris}
\usage{
death_data_format(
  date = NULL,
  deaths = NULL,
  cases = NULL,
  reporting_quality = 0.2
)
}
\arguments{
\item{date}{Character or Date vector of time series}

\item{deaths}{Numeric vector of deaths}

\item{cases}{Numeric vector of deaths}

\item{reporting_quality}{When generating synthetic data, what is the
reporting quality, i.e. probability of reporting a death prior to the current
total}
}
\value{
Time series of deaths as \code{data.frame}
}
\description{
Generate death time seris
}
\details{
Create a data frame for time series of
  deaths. If not provided, dummy data will
  be generated. The resulting data frame will be used
  when calibrating the model to fit to timeseries of
  deaths.

  In the future maybe extend this to include ITU cases
  and general hospital cases.
}
