% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{default_durations}
\alias{default_durations}
\title{Return the default hospital durations for modelling}
\usage{
default_durations()
}
\value{
List of default durations
}
\description{
Return the default hospital durations for modelling
}
\details{
\itemize{
   \item{tt_dur_get_ox_survive = 0}
   \item{tt_dur_get_mv_survive = 0}
   \item{dur_get_ox_survive = 9}
   \item{dur_get_ox_die = 9}
   \item{dur_not_get_ox_survive = 9 * 0.5}
   \item{dur_not_get_ox_die = 9 * 0.5}
   \item{dur_get_mv_survive = 15.3}
   \item{dur_get_mv_die = 11.5}
   \item{dur_not_get_mv_survive = 15.3 * 0.5}
   \item{dur_not_get_mv_die = 1}
   \item{dur_rec = 3}
}
}
