% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{default_probs}
\alias{default_probs}
\title{Return the default probabilities for modelling}
\usage{
default_probs()
}
\value{
list of default probabilities
}
\description{
Return the default probabilities for modelling
}
