% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{default_vaccine_pars}
\alias{default_vaccine_pars}
\title{Return the default vaccine parameters for modelling}
\usage{
default_vaccine_pars()
}
\value{
list of default vaccine parameters
}
\description{
Return the default vaccine parameters for modelling
}
