% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{describe_treatment_effects}
\alias{describe_treatment_effects}
\title{Describe Treatment Effects}
\usage{
describe_treatment_effects(param_values, param_name)
}
\arguments{
\item{param_values}{Parameter values for different treatments}

\item{param_name}{Parameter name}
}
\value{
Treatment effect description
}
\description{
Describe how treatments affect each parameter
}
