% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{determine_geometry_partitioning}
\alias{determine_geometry_partitioning}
\title{Determine Geometry Partitioning}
\usage{
determine_geometry_partitioning(data, response_type)
}
\arguments{
\item{data}{Experimental data frame}

\item{response_type}{Type of biological response}
}
\value{
Geometry configuration
}
\description{
Determine parameter types for geometry-adaptive optimization
}
