% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{df_output_wide}
\alias{df_output_wide}
\title{Create wide data.frame from odin output element}
\usage{
df_output_wide(m, compartment)
}
\arguments{
\item{m}{Model}

\item{compartment}{Output name (from odin output)}
}
\value{
Data.frame for variable
}
\description{
Create wide data.frame from odin output element
}
