% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elderly_pop}
\alias{elderly_pop}
\title{Elderly Population}
\format{
A data.frame
}
\usage{
elderly_pop
}
\description{
Population of 80-84, 85-89 and 90+ by age group and country
}
\keyword{datasets}
