% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{explicit_env_model}
\alias{explicit_env_model}
\title{Explicit environmental SEEIR model creation.

We will use this structure to ensure that model fitting is flexible in the
future as more models are added}
\usage{
explicit_env_model()
}
\description{
Create an explicit environmental model
}
