% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{explicit_model}
\alias{explicit_model}
\title{Explicit SEEIR model creation.

We will use this structure to ensure that model fitting is flexible in the
future as more models are added}
\usage{
explicit_model()
}
\description{
Create an explicit model
}
