% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extract_all_outputs}
\alias{extract_all_outputs}
\title{Extract All Model Ouputs}
\usage{
extract_all_outputs(x)
}
\arguments{
\item{x}{\code{squire_simulation} object representing the output from running
\code{run_explicit_SEEIR_model(output_transform = TRUE)}}
}
\value{
output
}
\description{
Extract All Model Ouputs
}
