% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{extract_deaths}
\alias{extract_deaths}
\title{Extract deaths from model output}
\usage{
extract_deaths(x, reduce_age = TRUE, data_0 = NULL)
}
\arguments{
\item{x}{squire_simulation object}

\item{reduce_age}{Collapse age-dimension}

\item{date_0}{Date of time 0, if specified a date column will be added}
}
\value{
Formatted long data.frame
}
\description{
Extract deaths from model output
}
