% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{extract_report_summaries}
\alias{extract_report_summaries}
\title{Extract report summaries}
\usage{
extract_report_summaries(x, date_0 = NULL)
}
\arguments{
\item{x}{squire_simulation object}

\item{date_0}{Date of time 0, if specified a date column will be added}
}
\value{
Formatted long data.frame
}
\description{
Extract report summaries
}
