% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extract_specific_output}
\alias{extract_specific_output}
\title{Extract Relevant Model Ouputs}
\usage{
extract_specific_output(squire_object, output_required)
}
\arguments{
\item{output_required}{The specific output you want out of the model ("infections", "deaths",
"hospital" or "ICU")}

\item{model_output}{squire_simulation object representing the output from running
run_explicit_SEEIR}
}
\value{
output
}
\description{
Extract Relevant Model Ouputs
}
