% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{fit_biological_model}
\alias{fit_biological_model}
\title{Fit Biological Model}
\usage{
fit_biological_model(data, response_type, galahad_config, verbose = FALSE)
}
\arguments{
\item{data}{Treatment-specific data frame}

\item{response_type}{Type of biological response}

\item{galahad_config}{GALAHAD configuration}

\item{verbose}{Print progress messages}
}
\value{
Model fitting results
}
\description{
Fit biological model to single treatment data
}
