% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{fit_germination_model}
\alias{fit_germination_model}
\title{Fit Germination Model}
\usage{
fit_germination_model(data, galahad_config, verbose = FALSE)
}
\arguments{
\item{data}{Germination data frame}

\item{galahad_config}{GALAHAD configuration}

\item{verbose}{Print progress messages}
}
\value{
Germination model results
}
\description{
Fit germination-specific model
}
