% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{format_deterministic_output}
\alias{format_deterministic_output}
\title{Format deterministic model output as data.frame}
\usage{
format_deterministic_output(x)
}
\arguments{
\item{x}{squire_simulation object}
}
\value{
Formatted long data.frame
}
\description{
Format deterministic model output as data.frame
}
