% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{format_output}
\alias{format_output}
\title{Format model output as data.frame}
\usage{
format_output(
  x,
  var_select = NULL,
  reduce_age = TRUE,
  reduce_compartment = TRUE
)
}
\arguments{
\item{x}{squire_simulation object}

\item{var_select}{Vector of variable names}

\item{reduce_age}{Collapse age-dimension}

\item{reduce_compartment}{Collapse compartments}
}
\value{
Formatted long data.frame
}
\description{
Format model output as data.frame
}
