% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{format_output_with_compartments}
\alias{format_output_with_compartments}
\title{Format model for general compartment structures}
\usage{
format_output_with_compartments(
  x,
  var_select = NULL,
  reduce_age = TRUE,
  combine_compartments = TRUE,
  date_0 = NULL,
  single_compartments,
  multi_compartments,
  summary_variable_compartments
)
}
\arguments{
\item{x}{squire_simulation object}

\item{var_select}{variables to select from
`summary_variable_compartments`}

\item{reduce_age}{Collapse age-dimension, calculating the total in the
compartment.}

\item{combine_compartments}{Collapse compartments of same type together
(e.g. E1 and E2 -> E)}

\item{date_0}{Date of time 0, if specified a date column will be added}

\item{single_compartments}{a vector of single compartments}

\item{multi_compartments}{a vector of linked compartments}

\item{summary_variable_compartments}{a mapping of variables to compartments}
}
\value{
Formatted long data.frame
}
\description{
Format model for general compartment structures
}
\details{
output with arbitrary compartment -> variable mappings as data.frame
}
