% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output_vaccine.R
\name{format_vaccine}
\alias{format_vaccine}
\title{Format vaccine model output as data.frame}
\usage{
format_vaccine(
  x,
  compartments = c("S", "E", "V", "EVac", "SVac", "RVac", "IMild", "ICase", "IICU",
    "IHospital", "IRec", "R", "D", "N"),
  summaries = c("deaths", "infections", "hospital_occupancy", "ICU_occupancy",
    "hospital_demand", "ICU_demand", "vaccines"),
  reduce_age = TRUE,
  date_0 = NULL
)
}
\arguments{
\item{x}{squire_simulation object}

\item{compartments}{Vector of compartment names, e.g. \code{c("S", "R")}.}

\item{summaries}{Vecto of summary names, which may be:
\itemize{
      \item{"deaths"}{ Daily Deaths (stochatic model) }
      \item{"infections"}{ Daily Infections. New infections (note this is currently a slightly different definitionto the main Squire mode)}
      \item{"hospital_occupancy"}{ Occupied Hospital Beds }
      \item{"ICU_occupancy"}{ Occupied ICU Beds }
      \item{"hospital_demand}{ Required Hospital Beds }
      \item{"ICU_demand}{ Required ICU Beds }
      \item{"vaccinated"}{ Daily vaccines administered (Stochastic model)}
      }}

\item{reduce_age}{Collapse age-dimension, calculating the total in the
compartment.}

\item{date_0}{Date of time 0, if specified a date column will be added}
}
\value{
Formatted long data.frame
}
\description{
Format vaccine model output as data.frame
}
