% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{generate_biological_interpretation}
\alias{generate_biological_interpretation}
\title{Generate Biological Interpretation}
\usage{
generate_biological_interpretation(
  optimization_results,
  parameter_validation,
  response_type
)
}
\arguments{
\item{optimization_results}{Optimization results}

\item{parameter_validation}{Parameter validation results}

\item{response_type}{Type of biological response}
}
\value{
Biological interpretation
}
\description{
Create biological interpretation of optimization results
}
