% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrices.R
\name{generate_contact_matrix}
\alias{generate_contact_matrix}
\title{Process a contact matrix}
\usage{
generate_contact_matrix(mixing_matrix, population)
}
\arguments{
\item{mixing_matrix}{Matrix}

\item{population}{Vector of population by age}
}
\value{
Processed matrix
}
\description{
Process a contact matrix
}
