% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{generate_recommendations}
\alias{generate_recommendations}
\title{Generate Recommendations}
\usage{
generate_recommendations(validation_results, parameter_validation)
}
\arguments{
\item{validation_results}{Statistical validation results}

\item{parameter_validation}{Parameter validation results}
}
\value{
Recommendations list
}
\description{
Provide methodological and experimental recommendations
}
