% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{get_biological_meaning}
\alias{get_biological_meaning}
\title{Get Biological Meaning}
\usage{
get_biological_meaning(param_name, response_type)
}
\arguments{
\item{param_name}{Parameter name}

\item{response_type}{Type of biological response}
}
\value{
Biological meaning description
}
\description{
Translate parameter names to biological interpretation
}
