% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{get_healthcare_capacity}
\alias{get_healthcare_capacity}
\title{Get healthcare capacity data}
\usage{
get_healthcare_capacity(country, simple_SEIR = FALSE)
}
\arguments{
\item{country}{Country name}

\item{simple_SEIR}{Logical. Is the population for the \code{simple_SEIR}.
Default = FALSE}
}
\value{
Healthcare capacity data
}
\description{
Get healthcare capacity data
}
