% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{get_mixing_matrix}
\alias{get_mixing_matrix}
\title{Get mixing matrix}
\usage{
get_mixing_matrix(country = NULL, iso3c = NULL)
}
\arguments{
\item{country}{Country name}

\item{iso3c}{ISO 3C Country Code}
}
\value{
Age mixing matrix
}
\description{
Get mixing matrix
}
