% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{income_strata_healthcare_capacity}
\alias{income_strata_healthcare_capacity}
\title{Income strata healthcare capacity}
\format{
A data.frame
}
\usage{
income_strata_healthcare_capacity
}
\description{
Average Hospital and ICU beds per 1000 pop for different
World Bank income groups
}
\keyword{datasets}
