% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{init_check}
\alias{init_check}
\title{Check and set up initial values}
\usage{
init_check(init, population)
}
\arguments{
\item{init}{Data.frame of initial conditions}

\item{population}{Population vector (for each age group)}
}
\value{
Checked initial values data.frame
}
\description{
Check and set up initial values
}
