% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{init_check_explicit}
\alias{init_check_explicit}
\title{Check and set up initial values for explict model}
\usage{
init_check_explicit(init, population, seeding_cases = 20)
}
\arguments{
\item{init}{Data.frame of initial conditions. Default = NULL}

\item{population}{Population vector (for each age group). Default = NULL,
which will cause population to be sourced from \code{country}}

\item{seeding_cases}{Initial number of cases seeding the epidemic}
}
\value{
Checked initial values data.frame
}
\description{
Check and set up initial values for explict model
}
