% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{long_output}
\alias{long_output}
\title{Convert model output to long format}
\usage{
long_output(m, vars = NULL)
}
\arguments{
\item{m}{Model output}

\item{vars}{Character vector of variable names to select}
}
\value{
Long format model output data.frame
}
\description{
Convert model output to long format
}
