% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{matrix_check}
\alias{matrix_check}
\title{Check dimensions of inputs}
\usage{
matrix_check(population, baseline_contact_matrix, contact_matrix_set)
}
\arguments{
\item{population}{Population vector (for each age group)}

\item{baseline_contact_matrix}{Contact matrix (in absence of interventions).
Used in estimation of beta from R0 vector}

\item{contact_matrix_set}{Contact matrices used in simulation}
}
\value{
Null if checks pass
}
\description{
Check dimensions of inputs
}
