% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{parse_country_severity}
\alias{parse_country_severity}
\title{Parse country severity parameters}
\usage{
parse_country_severity(
  country = NULL,
  prob_hosp = NULL,
  prob_severe = NULL,
  prob_non_severe_death_treatment = NULL,
  prob_severe_death_treatment = NULL,
  prob_non_severe_death_no_treatment = NULL,
  prob_severe_death_no_treatment = NULL,
  walker_params = FALSE
)
}
\arguments{
\item{country}{Character for country beign simulated. WIll be used to
generate \code{population} and \code{contact_matrix_set} if
unprovided. Either \code{country} or \code{population} and
\code{contact_matrix_set} must be provided.}

\item{prob_hosp}{Probability of hospitalisation by age.
Default, NULL, will use
\code{c(0.000840764, 0.001182411, 0.001662887, 0.002338607,
0.003288907, 0.004625365, 0.006504897, 0.009148183, 0.012865577,
0.018093546, 0.025445917, 0.035785947, 0.050327683, 0.0707785,
0.099539573, 0.1399878, 0.233470395)}}

\item{prob_severe}{Probability of developing severe symptoms by age.
Default, NULL, will use
\code{c(0.000840764, 0.001182411, 0.001662887, 0.002338607,
0.003288907, 0.004625365, 0.006504897, 0.009148183, 0.012865577,
0.018093546, 0.025445917, 0.035785947, 0.050327683, 0.0707785,
0.099539573, 0.1399878, 0.233470395)}}

\item{prob_non_severe_death_treatment}{Probability of death from non severe
treated infection. Default, NULL, will use
\code{c(0.181354223, 0.181354223, 0.181354223, 0.137454906,
0.121938236, 0.122775613, 0.136057441, 0.160922182, 0.196987378,
0.242011054, 0.289368845, 0.326537862, 0.337229819, 0.309082553,
0.243794865, 0.160480254, 0.057084366)}}

\item{prob_severe_death_treatment}{Probability of death from severe infection
that is treated. Default, NULL, will use
\code{c(0.226668959, 0.252420241, 0.281097009, 0.413005389,
0.518451493, 0.573413613, 0.576222065, 0.54253573, 0.493557696,
0.447376527, 0.416666608, 0.411186639, 0.443382594, 0.538718871,
0.570434076, 0.643352843, 0.992620047)}}

\item{prob_non_severe_death_no_treatment}{Probability of death in non severe
hospital inections that aren't treated. Default, NULL, will use
\code{rep(0.5, 17)}}

\item{prob_severe_death_no_treatment}{Probability of death from severe infection
that is not treated. Default, NULL, will use
\code{rep(0.95, 17)}}

\item{walker_params}{Boolean for using parameters in Walker et al. Default = FALSE,
which uses parameter update as of November 2020. For full information see
parameters vignette}
}
\description{
Parse country severity parameters
}
