% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle.R
\name{particle_filter}
\alias{particle_filter}
\title{Run a particle filter}
\usage{
particle_filter(
  data,
  model,
  compare,
  n_particles,
  forecast_days = 0,
  save_particles = FALSE,
  full_output = FALSE,
  save_sample_state = FALSE,
  save_end_states = FALSE
)
}
\arguments{
\item{data}{Data to fit to.  This must be constructed with
\code{particle_filter_data}}

\item{model}{An odin model, used to generate stochastic samples}

\item{compare}{A function to generate log-weights}

\item{n_particles}{Number of particles}

\item{forecast_days}{Number of days to forecast forward from end
states.  Requires that \code{save_particles} is \code{TRUE}.}

\item{save_particles}{Logical, indicating if we save full particle
histories (this is slower).}

\item{full_output}{Logical, indicating whether the full model output,
including the state and the declared outputs are returned. Deafult = FALSE}

\item{save_sample_state}{Logical, indicating whether we should save a
single particle, chosen at random, at the final time point for which
we have data}

\item{save_end_states}{Logical, indicating whether we should save all
particles at the final time point for which we have data}
}
\description{
Run a particle filter
}
