% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle.R
\name{particle_filter_data}
\alias{particle_filter_data}
\title{Prepare particle filter data}
\usage{
particle_filter_data(data, start_date, steps_per_day)
}
\arguments{
\item{data}{A data.frame of observed data.  There must be a column
\code{date}, containing dates (or ISO-formatted strings for
conversion with \code{\link{as.Date}}.}

\item{start_date}{The date to start the simulation from.  Must be
earlier than the first date in \code{data}.}

\item{steps_per_day}{The number of steps per day}
}
\description{
Prepare data for use with the particle filter.  This function
exists to make explicit how time changes through the model
relative to the data and to odin's internal clock.
}
