% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{perform_validated_optimization}
\alias{perform_validated_optimization}
\title{Perform Validated Optimization}
\usage{
perform_validated_optimization(
  data,
  treatments,
  galahad_config,
  validation_results,
  response_type,
  verbose = FALSE
)
}
\arguments{
\item{data}{Experimental data frame}

\item{treatments}{Treatment names vector}

\item{galahad_config}{GALAHAD configuration}

\item{validation_results}{Statistical validation results}

\item{response_type}{Type of biological response}

\item{verbose}{Print progress messages}
}
\value{
Optimization results
}
\description{
Execute parameter optimization with statistical validation
}
