% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.squire_calibration}
\alias{plot.squire_calibration}
\title{Calibrated Model PPlotting}
\usage{
\method{plot}{squire_calibration}(x, ..., what = "cases", forecast = 0)
}
\arguments{
\item{x}{\code{squire_calibration} object produced from
from \code{\link{calibrate_output_parsing}}}

\item{...}{Other parameters to be passed to internal plotting functions.}

\item{what}{What plotting type are we plotting. Options are \code{cases}
(default) or \code{healthcare}.}

\item{forecast}{How many days forward should forecast plots be provided.
Default = 0 days}
}
\value{
List of unformatted odin outputs with the date
}
\description{
Calibrated Model PPlotting
}
\details{
Fit the explicit_SEEIR model to time series of deaths
}
