% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{plot.squire_scan}
\alias{plot.squire_scan}
\title{squire scan plot}
\usage{
\method{plot}{squire_scan}(x, ..., what = "likelihood", log = FALSE)
}
\arguments{
\item{x}{An squire_scan object}

\item{...}{additional arguments affecting the plot produced.}

\item{what}{What scan outputs are we plotting of "likelihood" vs "probability"}

\item{log}{Should the axes by plotted on log scale}
}
\description{
squire scan plot
}
