% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{plot.squire_untransform}
\alias{plot.squire_untransform}
\title{squire simulation plot}
\usage{
\method{plot}{squire_untransform}(gcu)
}
\arguments{
\item{x}{A \code{squire_untransform} object from
\code{\link{untransformed_output}}}
}
\description{
squire simulation plot
}
