% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{pos_num}
\alias{pos_num}
\title{Check argument is a single positive numeric}
\usage{
pos_num(x, name)
}
\arguments{
\item{x}{argument}

\item{name}{Name of argument}
}
\value{
Nothing if check pass
}
\description{
Check argument is a single positive numeric
}
