% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{print.SQUIRE}
\alias{print.SQUIRE}
\title{Print SQUIRE Results}
\usage{
\method{print}{SQUIRE}(x, ...)
}
\arguments{
\item{x}{SQUIRE results object}

\item{...}{Additional arguments}
}
\description{
Print SQUIRE Results
}
