% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{print.squire_simulation}
\alias{print.squire_simulation}
\title{squire simulation print}
\usage{
\method{print}{squire_simulation}(x, ...)
}
\arguments{
\item{x}{An iccm_simulation object}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
squire simulation print
}
