% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrices.R
\name{process_contact_matrix_scaled_age}
\alias{process_contact_matrix_scaled_age}
\title{Process a contact matrix with an extra}
\usage{
process_contact_matrix_scaled_age(contact_matrix, population)
}
\arguments{
\item{contact_matrix}{A contact matrix}

\item{population}{Vector of population by age}
}
\value{
Processed matrix
}
\description{
Process a contact matrix with an extra
}
