% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{projection_plotting}
\alias{projection_plotting}
\title{Plot projections against each other}
\usage{
projection_plotting(
  r_list,
  scenarios,
  add_parms_to_scenarios = TRUE,
  var_select = NULL,
  replicates = FALSE,
  summarise = TRUE,
  ci = TRUE,
  q = c(0.025, 0.975),
  summary_f = mean,
  date_0 = Sys.Date(),
  x_var = "t",
  ...
)
}
\arguments{
\item{r_list}{List of different projection runs from \code{\link{projections}}}

\item{scenarios}{Character vector describing the different scenarios.}

\item{add_parms_to_scenarios}{Logical. Should the parameters used for the
projection runs be added to scenarios. Default = TRUE}

\item{var_select}{Vector of variable names to plot (default is all)}

\item{replicates}{Plot replicates}

\item{summarise}{Logical, add summary line}

\item{ci}{logical add confidence interval ribbon}

\item{q}{Quantiles for upper and lower of interval ribbon}

\item{summary_f}{Function to summarise each compartment
passed to the \code{fun} argument of \code{\link[ggplot2]{stat_summary}}}

\item{date_0}{Date of time 0, if specified a date column will be added}

\item{x_var}{X variable to use for plotting (default is \code{"t"},
but can be set to, \code{"date"}, if \code{date_0} provided), which will
cause the date to be plotted rather than time.}

\item{...}{additional arguments passed to \code{\link{format_output}}}
}
\description{
Plot projections against each other
}
