% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{projections}
\alias{projections}
\title{Provide projections from calibrated simulations}
\usage{
projections(r, mitigation, tt_mitigation)
}
\arguments{
\item{r}{Calibrated \code{\link{squire_simulation}} object.}

\item{mitigation}{Numeric vector for changes to R0 since calibration.}

\item{tt_mitigation}{Change time points for mitigation}
}
\description{
Provide projections from calibrated simulations
}
