% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggering.R
\name{projections_by_replicate}
\alias{projections_by_replicate}
\title{Provide projections from calibrated simulations by changing RO, contact
matrices or bed availability.}
\usage{
projections_by_replicate(
  r,
  time_period = NULL,
  R0 = NULL,
  R0_change = NULL,
  tt_R0 = 0,
  contact_matrix_set = NULL,
  contact_matrix_set_change = NULL,
  tt_contact_matrix = 0,
  hosp_bed_capacity = NULL,
  hosp_bed_capacity_change = NULL,
  tt_hosp_beds = 0,
  ICU_bed_capacity = NULL,
  ICU_bed_capacity_change = NULL,
  tt_ICU_beds = 0,
  to_be_run = TRUE
)
}
\arguments{
\item{r}{Calibrated \code{{squire_simulation}} object.}

\item{time_period}{How many days is the projection. Deafult = NULL, which will
carry the projection forward from t = 0 in the calibration (i.e. the number
of days set in calibrate using forecast)}

\item{R0}{Numeric vector for R0 from t = 0 in the calibration.
E.g. \code{R0 = c(2, 1)}. Default = NULL, which will use \code{R0_change}
to alter R0 if provided.}

\item{R0_change}{Numeric vector for relative changes in R0 relative to the
final R0 used in the calibration (i.e. at t = 0 in the calibration)
E.g. \code{R0 = c(0.8, 0.5)}. Default = NULL, which will use \code{R0} to
parameterise changes in R0 if provided.}

\item{tt_R0}{Change time points for R0}

\item{contact_matrix_set}{Contact matrices used in simulation. Default =
NULL, which will use \code{contact_matrix_set_change} to alter the contact
matrix if provided.}

\item{contact_matrix_set_change}{Numeric vector for relative changes in the
contact matrix realtive to the final contact matrix used in the calibration
(i.e. at t = 0 in the calibration).
E.g. \code{contact_matrix_set_change = c(0.8, 0.5)}. Default = NULL, which
will use \code{contact_matrix_set} to parameterise changes in contact
matrices if if provided.}

\item{tt_contact_matrix}{Time change points for matrix change. Default = 0}

\item{hosp_bed_capacity}{Numeric vector for hospital bed capacity
from t = 0 in the calibration. Default = NULL, which will use
\code{hosp_bed_capacity_change} to alter hosp_bed_capacity if provided.}

\item{hosp_bed_capacity_change}{Numeric vector for relative changes in
hospital bed capacity relative to the final hospital bed capacity used in the
calibration (i.e. at t = 0 in the calibration).
E.g. \code{hosp_bed_capacity_change = c(0.8, 0.5)}. Default = NULL, which
will use \code{hosp_bed_capacity} to parameterise changes in hospital bed capacity
if provided.}

\item{tt_hosp_beds}{Change time points for hosp_bed_capacity}

\item{ICU_bed_capacity}{Numeric vector for ICU bed capacity
from t = 0 in the calibration. Default = NULL, which will use
\code{ICU_bed_capacity_change} to alter ICU_bed_capacity if provided.}

\item{ICU_bed_capacity_change}{Numeric vector for relative changes in
ICU bed capacity relative to the final ICU bed capacity used in the
calibration (i.e. at t = 0 in the calibration).
E.g. \code{ICU_bed_capacity_change = c(0.8, 0.5)}. Default = NULL, which
will use \code{ICU_bed_capacity} to parameterise changes in ICU bed capacity
if provided.}

\item{tt_ICU_beds}{Change time points for ICU_bed_capacity}

\item{to_be_run}{List of logicals for whether each replicate should be run.
Default = TRUE, which causes all replictes to be run.}
}
\description{
This differs to \code{projections} as you can pass in lists of each argument
that then get passed to each simulation replicate. In time this will become
\code{projections} as it will have the same functionality.
}
\details{
The user can specify changes to R0, contact matrices and bed
provision, which will come into effect from the current day in the calibration.
These changes can either set these to be specific values or change them
relative to their values in the original simulation. If no change is
requested, the simulation will use parameters chosen for the calibration run.
}
