% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{quick_long}
\alias{quick_long}
\title{Convert squire_simulation object output to lond data.frame}
\usage{
quick_long(sim)
}
\arguments{
\item{sim}{squire_simulation object}
}
\value{
Long data.frame of output
}
\description{
Convert squire_simulation object output to lond data.frame
}
