% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle.R
\name{run_deterministic_comparison}
\alias{run_deterministic_comparison}
\title{Run Deterministic model comparison}
\usage{
run_deterministic_comparison(
  data,
  squire_model,
  model_params,
  model_start_date = "2020-02-02",
  obs_params = list(phi_cases = 0.1, k_cases = 2, phi_death = 1, k_death = 2, exp_noise =
    1e+06),
  forecast_days = 0,
  save_history = FALSE,
  return = "ll"
)
}
\arguments{
\item{data}{to fit to.}

\item{squire_model}{A squire model to use}

\item{model_params}{Squire model parameters. Created from a call to one of
the \code{parameters_<type>_model} functions.}

\item{model_start_date}{Date to run model simulations from}

\item{obs_params}{List of parameters used for comparing
model to data}

\item{forecast_days}{Days ahead to include in output}

\item{save_history}{Whether to save full history. Default = FALSE}

\item{return}{Set return depending on what is needed. 'full' and "sample" gives
the entire output, 'll' gives the log-likelihood}
}
\value{
Results from particle filter
}
\description{
Create a deterministic model and compare to data
}
