% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_explicit_SEEIR_model}
\alias{run_explicit_SEEIR_model}
\title{-----------------------------------------------------------------------------
Run the explicit SEEIR model}
\usage{
run_explicit_SEEIR_model(
  population,
  baseline_contact_matrix,
  contact_matrix_set,
  tt_contact_matrix = 0,
  R0 = 3,
  tt_R0 = 0,
  beta_set = NULL,
  time_period = 365,
  dt = 0.25,
  replicates = 10,
  init = NULL,
  prob_hosp = c(0.001127564, 0.000960857, 0.001774408, 0.003628171, 0.008100662,
    0.015590734, 0.024597885, 0.035377529, 0.04385549, 0.058495518, 0.08747709,
    0.109730508, 0.153943118, 0.177242143, 0.221362219, 0.267628264),
  prob_severe = c(3.73755e-05, 3.18497e-05, 5.88166e-05, 0.000120264, 0.000268514,
    0.000516788, 0.00081535, 0.001242525, 0.001729275, 0.002880196, 0.00598205,
    0.010821894, 0.022736324, 0.035911156, 0.056362032, 0.081467057),
  prob_non_severe_death = c(0.0125702, 0.0125702, 0.0125702, 0.0125702, 0.0125702,
    0.0125702, 0.0125702, 0.013361147, 0.015104687, 0.019164124, 0.027477519, 0.041762108,
    0.068531658, 0.105302319, 0.149305732, 0.20349534),
  prob_severe_death = c(0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5,
    0.5, 0.5, 0.5),
  dur_E = 4.58,
  dur_R = 2.09,
  dur_hosp = 5,
  dur_ox = 6,
  dur_mv = 5.5,
  dur_rec = 6
)
}
\arguments{
\item{population}{Population vector (for each age group).}

\item{baseline_contact_matrix}{Contact matrix (in absence of interventions).
Used in estimation of beta from R0 vector}

\item{contact_matrix_set}{Contact matrices used in simulation}

\item{tt_contact_matrix}{Time change points for matrix change. Default = 0}

\item{R0}{Basic Reproduction Number. Default = 3}

\item{tt_R0}{Change time points for R0. Default = 0}

\item{beta_set}{Alternative parameterisation via beta rather than R0.
Default = NULL, which causes beta to be estimated from R0}

\item{time_period}{Length of simulation. Default = 365}

\item{dt}{Time Step. Default = 0.25}

\item{replicates}{Number of replicates. Default = 10}

\item{init}{Data.frame of initial conditions. Default = NULL}

\item{prob_hosp}{probability of hospitalisation by age.
Default = c(0.001127564, 0.000960857, 0.001774408, 0.003628171,
0.008100662, 0.015590734, 0.024597885, 0.035377529,
0.04385549, 0.058495518, 0.08747709, 0.109730508,
0.153943118, 0.177242143, 0.221362219, 0.267628264)}

\item{prob_severe}{Probability of developing severe symptoms by age.
Default = c(3.73755e-05, 3.18497e-05, 5.88166e-05, 0.000120264,
0.000268514, 0.000516788, 0.00081535, 0.001242525,
0.001729275, 0.002880196, 0.00598205, 0.010821894,
0.022736324, 0.035911156, 0.056362032, 0.081467057)}

\item{prob_non_severe_death}{Probability of death from non severe infection.
Default = c(0.0125702, 0.0125702, 0.0125702, 0.0125702,
0.0125702, 0.0125702, 0.0125702, 0.013361147,
0.015104687, 0.019164124, 0.027477519, 0.041762108,
0.068531658, 0.105302319, 0.149305732, 0.20349534)}

\item{prob_severe_death}{Probability of death from severe infection.
Default = c(0.5, 0.5, 0.5, 0.5,
0.5, 0.5, 0.5, 0.5,
0.5, 0.5, 0.5, 0.5,
0.5, 0.5, 0.5, 0.5)}

\item{dur_E}{Mean duration of incubation period (days). Default = 4.58}

\item{dur_R}{Mean duration of mild infection (days). Default = 2.09}

\item{dur_hosp}{Mean duration of hospitilsation (days). Default = 5}

\item{dur_ox}{Mean duration of case requiring oxygen (days). Default = 6}

\item{dur_mv}{Mean duration of mechanical ventilation (days). Default = 5.5}

\item{dur_rec}{Mean duration of case recovering after ICU (days).  Default = 6}
}
\value{
Simulation output
}
\description{
-----------------------------------------------------------------------------
Run the explicit SEEIR model
}
\examples{
\dontrun{
pop <- get_population("Afghanistan")
m1 <- run_explicit_SEEIR_model(R0 = 2,
population = pop$n, dt = 1,
baseline_contact_matrix = contact_matrices[[1]],
contact_matrix_set=contact_matrices[[1]])
}
}
