% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_simple_SEEIR_model}
\alias{run_simple_SEEIR_model}
\title{Run the SEEIR model}
\usage{
run_simple_SEEIR_model(
  R0 = 3,
  tt_R0 = 0,
  dt = 0.1,
  init = NULL,
  dur_E = 4.58,
  dur_I = 2.09,
  day_return = FALSE,
  population,
  contact_matrix_set,
  tt_contact_matrix = 0,
  time_period = 365,
  replicates = 10
)
}
\arguments{
\item{R0}{Basic reproduction number}

\item{tt_R0}{Change time points for R0}

\item{dt}{Time step}

\item{init}{Data.frame of initial conditions}

\item{dur_E}{Mean duration of incubation period (days)}

\item{dur_I}{Mean duration of infectious period (days)}

\item{day_return}{Logical, do we want to return outut after each day rather
than each dt. Default = FALSE}

\item{population}{Population vector (for each age group)}

\item{contact_matrix_set}{Contact matrices used in simulation}

\item{tt_contact_matrix}{Time change points for matrix change}

\item{time_period}{Length of simulation}

\item{replicates}{Number of replicates}
}
\value{
Simulation output
}
\description{
Run the SEEIR model
}
\examples{
\dontrun{
pop <- get_population("Afghanistan", simple_SEIR = TRUE)
m1 <- run_simple_SEEIR_model(population = pop$n, dt = 1,
R0 = 2,
contact_matrix_set=contact_matrices[[1]])
}
}
