% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{sample_3d_grid_scan}
\alias{sample_3d_grid_scan}
\title{Sample Grid Scan}
\usage{
sample_3d_grid_scan(
  scan_results,
  n_sample_pairs = 10,
  n_particles = 100,
  forecast_days = 0,
  full_output = FALSE
)
}
\arguments{
\item{scan_results}{Output of \code{\link{scan_R0_date_Meff}}.}

\item{n_sample_pairs}{Number of parameter pairs to be sampled. This will
determine how many trajectories are returned. Integer. Default = 10. This
will determine how many trajectories are returned.}

\item{n_particles}{Number of particles. Positive Integer. Default = 100}

\item{forecast_days}{Number of days being forecast. Default = 0}

\item{full_output}{Logical, indicating whether the full model output,
including the state and the declared outputs are returned. Deafult = FALSE}
}
\value{
\code{\link{list}}. First element (trajectories) is a 3
  dimensional array of trajectories (time, state, tranjectories). Second
  element (param_grid) is the parameters chosen when  sampling from the
  \code{scan_results} grid and the third dimension (inputs) is a list of
  model inputs.
}
\description{
Take a grid search produced by \code{\link{scan_R0_date_Meff}} and
sample \code{n_sample_pairs} from the parameter grid uses based
on their probability. For each parameter pair chosen, run particle
filter with \code{num_particles} and sample 1 trajectory
}
