% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc_utils.R
\name{sample_pmcmc}
\alias{sample_pmcmc}
\title{Sample PMCMC}
\usage{
sample_pmcmc(
  pmcmc_results,
  burnin = 0,
  n_chains,
  n_trajectories = 10,
  n_particles = 100,
  forecast_days = 0
)
}
\arguments{
\item{pmcmc_results}{output of \code{\link{run_mcmc_chain}}; The results from the PMCMC run -- can have mutliple chains.}

\item{burnin}{integer; Number of iterations to discard from the start of MCMC run. Default = 0}

\item{n_trajectories}{interger; Number of trajectories to be returned. Integer. Default = 10.}

\item{n_particles}{integer; Number of particles to be considered in the particle filter. Default = 100}

\item{forecast_days}{integer; number of days being forecast. Default = 0}

\item{full_output}{logical; Indicator for whether the full model output,
including the state and the declared outputs are returned. Deafult = FALSE}
}
\value{
\code{\link{list}}. First element (trajectories) is a 3
  dimensional array of trajectories (time, state, tranjectories). Second
  element (sampled_PMCMC_Results) is the parameters chosen when sampling from the
  \code{pmcmc} and the third dimension (inputs) is a list of
  model inputs.
}
\description{
Sample from the posterior probability results produced by \code{\link{run_mcmc_chain}}
to select parameter set. For each parmater set sampled, run particle
filter with \code{num_particles} and sample 1 trajectory
}
