% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{simple_model}
\alias{simple_model}
\title{Simple SEEIR model creation.

We will use this structure to ensure that model fitting is flexible in the
future as more models are added}
\usage{
simple_model()
}
\description{
Create a simple  model
}
