% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{summarize_biological_effects}
\alias{summarize_biological_effects}
\title{Summarize Biological Effects}
\usage{
summarize_biological_effects(interpretation, response_type)
}
\arguments{
\item{interpretation}{Parameter interpretation results}

\item{response_type}{Type of biological response}
}
\value{
Biological summary text
}
\description{
Create overall biological summary
}
