% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{summary.SQUIRE}
\alias{summary.SQUIRE}
\title{SQUIRE Summary Method}
\usage{
\method{summary}{SQUIRE}(object, ...)
}
\arguments{
\item{object}{SQUIRE results object}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects (prints summary to console)
}
\description{
Print method for SQUIRE results
}
