% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{summary.squire_simulation}
\alias{summary.squire_simulation}
\title{squire simulation summary}
\usage{
\method{summary}{squire_simulation}(object, ...)
}
\arguments{
\item{object}{An squire_simulation object}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
squire simulation summary
}
