% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{test_treatment_effects}
\alias{test_treatment_effects}
\title{Test Treatment Effects}
\usage{
test_treatment_effects(data, treatments, alpha = 0.05, verbose = FALSE)
}
\arguments{
\item{data}{Experimental data frame}

\item{treatments}{Treatment names vector}

\item{alpha}{Significance level}

\item{verbose}{Print progress messages}
}
\value{
List with statistical test results
}
\description{
Statistical test for treatment differences
}
