% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggering.R
\name{trigger_projections}
\alias{trigger_projections}
\title{Project lockdowns based on triggering}
\usage{
trigger_projections(
  out,
  trigger_metric = "deaths",
  trigger_value = 150,
  R0_lockdowns = c(0.5, 0.5, 0.5, 0.5),
  lockdown_lengths = c(28, 42, 28, 42),
  max_lockdowns = 4,
  seed = 931L
)
}
\arguments{
\item{out}{Output of \code{\link{pmcmc}} or \code{\link{calibrate}}}

\item{trigger_metric}{Name of model output to tigger by. Must be one accepted
by \code{format_output}}

\item{trigger_value}{Value of \code{trigger_metric} at which triggering occurs}

\item{R0_lockdowns}{Vector of R0 values to be used for each lockdown.
\code{Default = c(0.5, 0.5, 0.5, 0.5)}}

\item{lockdown_lengths}{Vector of lengths of each lockdown in days.
\code{Default = c(28, 42, 28, 42)}}

\item{max_lockdowns}{Maximum number of lockdowns. \code{Default = 4}}

\item{seed}{RNG seed to be used. \code{Default = 931L}}
}
\description{
Project lockdowns based on triggering
}
