% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigger_running.R
\name{trigger_run}
\alias{trigger_run}
\title{Trigger run}
\usage{
trigger_run(
  trigger_type = NULL,
  trigger_threshold = NULL,
  replicates = 25,
  country = NULL,
  population = NULL,
  contact_matrix_set = NULL,
  ...
)
}
\arguments{
\item{trigger_type}{Type of trigger (deaths, ICU capacity)}

\item{trigger_threshold}{Number of observed deaths suppression to be triggered at}

\item{replicates}{Number of replicates. Default = 10}

\item{country}{Character for country beign simulated. WIll be used to
generate \code{population} and \code{contact_matrix_set} if
unprovided. Either \code{country} or \code{population} and
\code{contact_matrix_set} must be provided.}

\item{population}{Population vector (for each age group). Default = NULL,
which will cause population to be sourced from \code{country}}

\item{contact_matrix_set}{Contact matrices used in simulation. Default =
NULL, which will generate this based on the \code{country}.}

\item{...}{Further aguments for \code{run_explicit_SEEIR_model()}}
}
\value{
List of time adjusted squire_simulations
}
\description{
Trigger run
}
