% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{untransformed_output}
\alias{untransformed_output}
\title{Explicit Output from non transformed run}
\usage{
untransformed_output(
  r,
  compartments = NULL,
  deaths = TRUE,
  cases = TRUE,
  beds = TRUE
)
}
\arguments{
\item{r}{Output of \code{\link{run_explicit_SEEIR_model}} with
\code{output_transform} is FALSE}

\item{compartments}{Specific compartments to extract. Default = NULL}

\item{deaths}{Should incident deaths be include. Default = TRUE}

\item{cases}{Should incident cases be include. Default = TRUE}

\item{beds}{Should ICU bed and hospital bed demand be include. Default = TRUE}
}
\value{
Data frame of case numbers, hospital beds, icu beds and deaths
}
\description{
Explicit Output from non transformed run
}
\details{
Parse output of \code{\link{run_explicit_SEEIR_model}} when
\code{output_transform} is FALSE
}
