% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc_utils.R
\name{update_covariance_matrix}
\alias{update_covariance_matrix}
\title{update_cov}
\usage{
update_covariance_matrix(
  covariance_matrix,
  i,
  mean_vector,
  current_parameters,
  number_of_parameters
)
}
\arguments{
\item{covariance_matrix}{the current covariance matrix being used to propose new parameter values}

\item{i}{the number of iterations since covariance matrix adaptation began}

\item{mean_vector}{initially this the mean of the columns of the MCMC chain. sequentially updated
by this function.}

\item{current_parameters}{the current parameter values in the MCMC chain}

\item{number_of_parameters}{the number of parameters currently being estimated}
}
\description{
Involved in the Robbins-Munro optimisation within the Metropolis-Hastings MCMC.
Function to iteratively update the covariance matrix used to propose new parameter
values.
}
