% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc_utils.R
\name{update_scaling_factor}
\alias{update_scaling_factor}
\title{update_sigma}
\usage{
update_scaling_factor(
  scaling_factor,
  acceptance,
  probability,
  i,
  number_of_parameters
)
}
\arguments{
\item{scaling_factor}{the scaling factor involved in the Robbins-Munro optimisation}

\item{acceptance}{whether or not the most recent parameter proposal was accepted}

\item{probability}{the required acceptance probability}

\item{i}{the iteration number}

\item{number_of_parameters}{the number of parameters currently being estimated}
}
\description{
Involved in the Robbins-Munro optimisation within the Metropolis-Hastings MCMC.
Function to iteratively update the scaling factor of the covariance matrix.
}
