% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{validate_biological_effects}
\alias{validate_biological_effects}
\title{Validate Biological Effects}
\usage{
validate_biological_effects(
  data,
  treatments,
  control_treatment,
  response_type,
  alpha,
  min_timepoints,
  min_replicates,
  verbose
)
}
\arguments{
\item{data}{Data frame with experimental data}

\item{treatments}{Vector of treatment names}

\item{control_treatment}{Name of control treatment}

\item{response_type}{Type of biological response}

\item{alpha}{Statistical significance level}

\item{min_timepoints}{Minimum required timepoints}

\item{min_replicates}{Minimum required replicates}

\item{verbose}{Print progress messages}
}
\value{
List with validation results
}
\description{
Test for statistically significant treatment effects before optimization
}
