% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIRE.R
\name{validate_optimized_parameters}
\alias{validate_optimized_parameters}
\title{Validate Optimized Parameters}
\usage{
validate_optimized_parameters(
  optimization_results,
  treatments,
  control_treatment,
  alpha = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{optimization_results}{Results from optimization}

\item{treatments}{Vector of treatment names}

\item{control_treatment}{Name of control treatment}

\item{alpha}{Significance level}

\item{verbose}{Print progress messages}
}
\value{
Parameter validation results
}
\description{
Test statistical significance of optimized parameters
}
