% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{wide_output}
\alias{wide_output}
\title{Convert model output to long format}
\usage{
wide_output(m)
}
\arguments{
\item{m}{Model output}
}
\value{
Long format model output data.frame
}
\description{
Convert model output to long format
}
