% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate deaths}
\usage{
calibrate(
  country,
  deaths,
  reporting_fraction = 1,
  seeding_age_groups = c("35-40", "40-45", "45-50", "50-55"),
  min_seeding_cases = 5,
  max_seeding_cases = 50,
  replicates = 100,
  dt = 0.1,
  ...
)
}
\arguments{
\item{country}{Country name}

\item{deaths}{Number of observed deaths}

\item{reporting_fraction}{REporting fraction}

\item{seeding_age_groups}{Age groups for seeding}

\item{min_seeding_cases}{Minimum seeding cases}

\item{max_seeding_cases}{Maximum seeding cases}

\item{replicates}{Replicates}

\item{dt}{dt}

\item{...}{Further aguments for \code{run_explicit_SEEIR_model()}}
}
\value{
List of time adjusted squire_simulations
}
\description{
Calibrate deaths
}
