% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate deaths}
\usage{
calibrate(
  country,
  deaths,
  reporting_fraction = 1,
  seeding_age_groups = c("35-40", "40-45", "45-50", "50-55"),
  min_seeding_cases = 5,
  max_seeding_cases = 50,
  R0 = c(2.7, 3, 3.5),
  replicates = 100,
  dt = 0.1,
  ...
)
}
\arguments{
\item{country}{Country name}

\item{deaths}{Number of observed deaths}

\item{reporting_fraction}{Reporting fraction. Numeric for what proportion of
the total deaths the reported deaths represent. E.g. 0.5 results in
the model calibrating to twice the deaths provided by \code{deaths}}

\item{seeding_age_groups}{Age groups for seeding}

\item{min_seeding_cases}{Minimum seeding cases}

\item{max_seeding_cases}{Maximum seeding cases}

\item{R0}{Vector or R0 values to sample from to introduce uncertainty
in predictions. Default = c(2.7, 3.0, 3.5)}

\item{replicates}{Replicates}

\item{dt}{dt}

\item{...}{Further aguments for \code{run_explicit_SEEIR_model()}}
}
\value{
List of time adjusted squire_simulations
}
\description{
Calibrate deaths
}
