% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate Model}
\usage{
calibrate(
  data,
  country,
  reporting_fraction = 1,
  seeding_age_groups = c("35-40", "40-45", "45-50", "50-55"),
  min_seeding_cases = 5,
  max_seeding_cases = 50,
  parse_output = TRUE,
  replicates = 100,
  dt = 0.5,
  ...
)
}
\arguments{
\item{data}{Data frame with 2 variables: date and deaths}

\item{country}{Character. Country data originates from.}

\item{reporting_fraction}{Numbeic. Fraction of deaths expected to have been reported. DEFAULT = 1}

\item{seeding_age_groups}{Character vector. Age groups seeding cases should be distributed into.}

\item{min_seeding_cases}{Numeric. Minimum number of seeding cases. DEFAULT = 5.}

\item{max_seeding_cases}{Numeric. Maximum number of seeding cases. DEFAULT = 50.}

\item{parse_output}{Logical. Should output be parsed ready for plotting.
Default = TRUE}

\item{replicates}{Simulation Repetitions. Default = 10}

\item{dt}{Time Step. Default = 0.25}

\item{...}{Other parameters to pass to \code{\link{run_explicit_SEEIR_model}}}
}
\value{
List of formatted odin outputs, the data it is calibrated to and
  the parameter set used in calibration
}
\description{
Calibrate Model
}
\details{
Fit the explicit_SEEIR model to time series of deaths
}
