% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate via particle filter grid search using time series of deaths}
\usage{
calibrate(
  data,
  R0_min,
  R0_max,
  R0_step,
  first_start_date,
  last_start_date,
  day_step,
  Meff_min = 1,
  Meff_max = 1,
  Meff_step = 0.1,
  squire_model = explicit_model(),
  pars_obs = NULL,
  forecast = 0,
  n_particles = 100,
  reporting_fraction = 1,
  replicates = 100,
  date_R0_change = NULL,
  R0_change = NULL,
  date_ICU_bed_capacity_change = NULL,
  baseline_ICU_bed_capacity = NULL,
  ICU_bed_capacity = NULL,
  date_hosp_bed_capacity_change = NULL,
  baseline_hosp_bed_capacity = NULL,
  hosp_bed_capacity = NULL,
  date_contact_matrix_set_change = NULL,
  baseline_contact_matrix = NULL,
  contact_matrix_set = NULL,
  country = NULL,
  population = NULL,
  dt = 0.1,
  ...
)
}
\arguments{
\item{data}{Deaths data to fit to. See \code{example_deaths.csv}
and \code{particle_filter_data()}}

\item{R0_min}{Minimum value of R0 in the search}

\item{R0_max}{Maximum value of R0 in the search}

\item{R0_step}{Step to increment R0 between min and max}

\item{first_start_date}{Earliest start date as 'yyyy-mm-dd'}

\item{last_start_date}{Latest start date as 'yyyy-mm-dd'}

\item{day_step}{Step to increment date in days}

\item{squire_model}{A squire model. Default = \code{explicit_SEIR()}}

\item{pars_obs}{list of parameters to use for the comparison function.}

\item{forecast}{Number of days to forecast forward. Default = 0}

\item{n_particles}{Number of particles. Positive Integer. Default = 100}

\item{reporting_fraction}{Reporting fraction. Numeric for what proportion of
the total deaths the reported deaths represent. E.g. 0.5 results in
the model calibrating to twice the deaths provided by \code{data$deaths}}

\item{replicates}{Replicates to be run. Default = 100}

\item{date_R0_change}{Calendar dates at which R0_change occurs.
Defaut = NULL, i.e. no change in R0}

\item{R0_change}{Numeric vector for relative changes in R0. Default = NULL,
i.e. no change in R0}

\item{date_ICU_bed_capacity_change}{Calendar dates at which ICU bed
capacity changes set in \code{model_params} change.
Defaut = NULL, i.e. no change}

\item{baseline_ICU_bed_capacity}{The starting number of ICU beds before
the epidemic started. Default = NULL, which will use the hospital beds data
for the country provided. If no country is provided then this is 3/100 of
hospital beds}

\item{ICU_bed_capacity}{Number of ICU beds at each date specified in
\code{date_ICU_bed_capacity_change}. Must be same length as
\code{date_ICU_bed_capacity_change}.}

\item{date_hosp_bed_capacity_change}{Calendar dates at which hospital bed
capacity changes set in \code{model_params} change.
Defaut = NULL, i.e. no change}

\item{baseline_hosp_bed_capacity}{The starting number of hospital beds before
the epidemic started. Default = NULL, which will use the hospital beds data
for the country provided. If no country is provided then this is 5/1000 of
the population}

\item{hosp_bed_capacity}{Number of hospital beds at each date specified in
\code{date_hosp_bed_capacity_change}. Must be same length as
\code{date_hosp_bed_capacity_change}.}

\item{date_contact_matrix_set_change}{Calendar dates at which the contact matrices
set in \code{model_params} change. Defaut = NULL, i.e. no change}

\item{baseline_contact_matrix}{The starting contact matrix prior to any changes
due to interventions or otherwise. Default = NULL, which will use the contact
matrix associated with the coutnry provided.}

\item{contact_matrix_set}{List of contact matrices to be used from the dates
provided in \code{date_contact_matrix_set_change}.Must be same length as
\code{date_contact_matrix_set_change}}

\item{country}{Character for country beign simulated. WIll be used to
generate \code{population} and \code{contact_matrix_set} if
unprovided. Either \code{country} or \code{population} and
\code{contact_matrix_set} must be provided.}

\item{population}{Population vector (for each age group). Default = NULL,
which will cause population to be sourced from \code{country}}

\item{dt}{Time Step. Default = 0.1}

\item{...}{Further aguments for the model parameter function. If using the
\code{\link{explicit_model}} (default) this will be
\code{parameters_explicit_SEEIR}.}
}
\value{
List of dated squire simulations
}
\description{
Calibrate via particle filter grid search using time series of deaths
}
