% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate Model}
\usage{
calibrate(data, country, parse_output = TRUE, replicates = 10, dt = 0.25, ...)
}
\arguments{
\item{data}{Data frame with 2 variables: date and deaths}

\item{country}{Character. Country data originates from.}

\item{parse_output}{Logical. Should output be parsed ready for plotting.
Default = TRUE}

\item{replicates}{Simulation Repetitions. Default = 10}

\item{dt}{Time Step. Default = 0.25}

\item{...}{Other parameters to pass to \code{\link{run_explicit_SEEIR_model}}}
}
\value{
List of formatted odin outputs, the data it is calibrated to and
  the parameter set used in calibration
}
\description{
Calibrate Model
}
\details{
Fit the explicit_SEEIR model to time series of deaths
}
