% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_output_parsing}
\alias{calibrate_output_parsing}
\title{Format output of calibration for plotting}
\usage{
calibrate_output_parsing(r)
}
\arguments{
\item{r}{Output of \code{\link{calibrate}}}
}
\value{
\code{list} with:
\itemize{
      \item{"df"}{data frame of case numbers, hospital beds, icu beds and deaths }
      \item{"data"}{raw data used in calibration}
      \item{"parameters"}{parameters used in simulation}
      }
}
\description{
Format output of calibration for plotting
}
\details{
Calibration output is taken to give time series of infections, cases,
cases requiring hospitilisation, case requiring critical care facilities. Used
in plotting for nowcasting reports.
}
