% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderly_plotting.R
\name{calibrate_output_parsing}
\alias{calibrate_output_parsing}
\title{Format output of calibration for plotting}
\usage{
calibrate_output_parsing(r, date_0 = Sys.Date())
}
\arguments{
\item{r}{Output of \code{\link{calibrate}}}

\item{date_0}{Date of time 0, if specified a date column will be added}
}
\value{
\code{list} with:
\itemize{
      \item{df:}{ Data frame of case numbers, hospital beds, ICU beds and deaths }
      \item{data:}{ Raw data used in calibration}
      \item{parameters:}{ Parameters used in simulation}
      }
}
\description{
Format output of calibration for plotting
}
\details{
Calibration output is taken to give time series of infections, cases,
cases requiring hospitilisation, case requiring critical care facilities. Used
in plotting for nowcasting reports.
}
