% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extract_all_outputs}
\alias{extract_all_outputs}
\title{Extract All Model Ouputs}
\usage{
extract_all_outputs(squire_object)
}
\arguments{
\item{model_output}{squire_simulation object representing the output from running
run_explicit_SEEIR}
}
\value{
output
}
\description{
Extract All Model Ouputs
}
