% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extract_specific_output}
\alias{extract_specific_output}
\title{Extract Relevant Model Ouputs}
\usage{
extract_specific_output(x, output_required)
}
\arguments{
\item{x}{\code{squire_simulation} object representing the output from running
\code{run_explicit_SEEIR_model(output_transform = TRUE)}}

\item{output_required}{Character for specifc output you want out of the model
c("infections", "deaths", "hospital" or "ICU")}
}
\value{
output
}
\description{
Extract Relevant Model Ouputs
}
