% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output_vaccine.R
\name{format_vaccine}
\alias{format_vaccine}
\title{Format vaccine model output as data.frame}
\usage{
format_vaccine(
  x,
  compartments = c("S", "E", "V", "E_vac", "IMild", "ICase", "IICU", "IHospital",
    "IRec", "R", "D"),
  summaries = c("infections", "deaths", "vaccines", "ICU_occupancy", "ICU_demand",
    "hospital_occupancy", "hospital_demand"),
  reduce_age = TRUE,
  date_0 = NULL
)
}
\arguments{
\item{x}{squire_simulation object}

\item{compartments}{Vector of compartment names, e.g. \code{c("S", "R")}.}

\item{summaries}{Vecto of summary names, which may be:
\itemize{
      \item{"deaths"}{ Daily Deaths }
      \item{"infections"}{ Daily Infections }
      \item{"hospital_occupancy"}{ Occupied Hospital Beds }
      \item{"ICU_occupancy"}{ Occupied ICU Beds }
      \item{"hospital_demand}{ Required Hospital Beds }
      \item{"ICU_demand}{ Required ICU Beds }
      \item{"vaccines}{ Daily vaccines administered}
      }}

\item{reduce_age}{Collapse age-dimension, calculating the total in the
compartment.}

\item{date_0}{Date of time 0, if specified a date column will be added}
}
\value{
Formatted long data.frame
}
\description{
Format vaccine model output as data.frame
}
