% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output_vaccine.R
\name{format_vaccine_internal}
\alias{format_vaccine_internal}
\title{Internals of Format vaccine model output as data.frame}
\usage{
format_vaccine_internal(
  x,
  compartments,
  summaries,
  reduce_age,
  index,
  time,
  replicate
)
}
\arguments{
\item{x}{squire_simulation object}

\item{compartments}{Vector of compartment names, e.g. \code{c("S", "R")}.}

\item{summaries}{Vecto of summary names, which may be:
\itemize{
      \item{"deaths"}{ Daily Deaths (stochatic model) }
      \item{"infections"}{ Daily Infections (stochastic model)}
      \item{"hospital_occupancy"}{ Occupied Hospital Beds }
      \item{"ICU_occupancy"}{ Occupied ICU Beds }
      \item{"hospital_demand}{ Required Hospital Beds }
      \item{"ICU_demand}{ Required ICU Beds }
      \item{"vaccinated"}{ Daily vaccines administered (Stochastic model)}
      }}

\item{reduce_age}{Collapse age-dimension, calculating the total in the
compartment.}

\item{index}{odin ouput index}

\item{time}{time vector}

\item{replicate}{outpu replicate number}
}
\description{
Internals of Format vaccine model output as data.frame
}
