% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{get_population}
\alias{get_population}
\title{Get population data}
\usage{
get_population(country)
}
\arguments{
\item{country}{Country name}
}
\value{
Population data.frame
}
\description{
Get population data
}
