% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrices.R
\name{matrix_set}
\alias{matrix_set}
\title{Process set of contact matrices}
\usage{
matrix_set(contact_matrix_set, population)
}
\arguments{
\item{contact_matrix_set}{Set of contact matrices}

\item{population}{Vector of populaion by age}
}
\value{
Processed set of contact matrices
}
\description{
Process set of contact matrices
}
