% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.squire_calibration}
\alias{plot.squire_calibration}
\title{Calibrate Model}
\usage{
\method{plot}{squire_calibration}(squire_calibration, what = "cases")
}
\arguments{
\item{squire_calibration}{\code{squire_calibration} object produced from
from \code{\link{calibrate_output_parsing}}}

\item{what}{What plotting type are we plotting. Options are \code{cases}
(default) or \code{healthcare}.}
}
\value{
List of unformatted odin outputs with the date
}
\description{
Calibrate Model
}
\details{
Fit the explicit_SEEIR model to time series of deaths
}
