% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{plot.squire_simulation}
\alias{plot.squire_simulation}
\title{squire simulation plot}
\usage{
\method{plot}{squire_simulation}(x, ...)
}
\arguments{
\item{x}{An iccm_simulation object}

\item{...}{additional arguments affecting the plot produced.}
}
\description{
squire simulation plot
}
