% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_deterministic_SEIR_model}
\alias{run_deterministic_SEIR_model}
\title{Run the deterministic explicit SEIR model}
\usage{
run_deterministic_SEIR_model(
  population,
  contact_matrix,
  tt_R0,
  R0_set,
  time_period,
  hosp_bed_capacity,
  ICU_bed_capacity
)
}
\arguments{
\item{population}{Population vector (for each age group). Default = NULL,
which will cause population to be sourced from \code{country}}

\item{contact_matrix}{Contact matrix to use in the simulation.}

\item{tt_R0}{Time change points for R0}

\item{R0_set}{The values of R0 to use for the simulation}

\item{time_period}{Length of simulation. Default = 365}

\item{hosp_bed_capacity}{General bed capacity. Can be single number of vector if capacity time-varies.}

\item{ICU_bed_capacity}{ICU bed capacity. Can be single number of vector if capacity time-varies.}
}
\value{
Simulation output
}
\description{
Run the deterministic explicit SEIR model
}
\examples{
\dontrun{
pop <- get_population("Afghanistan")
m <- get_mixing_matrix("Afghanistan")
run_deterministic_SEIR_model(pop$n, m, [0, 50], [3, 3/2], 365, 100000,
1000000)
}
}
