% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_explicit_SEEIR_model}
\alias{run_explicit_SEEIR_model}
\title{Run the explicit SEEIR model}
\usage{
run_explicit_SEEIR_model(
  country = NULL,
  population = NULL,
  tt_contact_matrix = 0,
  contact_matrix_set = NULL,
  R0 = 3,
  tt_R0 = 0,
  beta_set = NULL,
  time_period = 365,
  dt = 0.1,
  day_return = FALSE,
  replicates = 10,
  init = NULL,
  seed = stats::runif(1, 0, 1e+08),
  prob_hosp = NULL,
  prob_severe = NULL,
  prob_non_severe_death_treatment = NULL,
  prob_non_severe_death_no_treatment = NULL,
  prob_severe_death_treatment = NULL,
  prob_severe_death_no_treatment = NULL,
  p_dist = probs$p_dist,
  walker = FALSE,
  dur_E = 4.6,
  dur_IMild = 2.1,
  dur_ICase = 4.5,
  dur_get_ox_survive = NULL,
  tt_dur_get_ox_survive = NULL,
  dur_get_ox_die = NULL,
  dur_not_get_ox_survive = NULL,
  dur_not_get_ox_die = NULL,
  dur_get_mv_survive = NULL,
  tt_dur_get_mv_survive = NULL,
  dur_get_mv_die = NULL,
  dur_not_get_mv_survive = NULL,
  dur_not_get_mv_die = NULL,
  dur_rec = NULL,
  hosp_bed_capacity = NULL,
  ICU_bed_capacity = NULL,
  tt_hosp_beds = 0,
  tt_ICU_beds = 0,
  seeding_cases = NULL
)
}
\arguments{
\item{country}{Character for country beign simulated. WIll be used to
generate \code{population} and \code{contact_matrix_set} if
unprovided. Either \code{country} or \code{population} and
\code{contact_matrix_set} must be provided.}

\item{population}{Population vector (for each age group). Default = NULL,
which will cause population to be sourced from \code{country}}

\item{tt_contact_matrix}{Time change points for matrix change. Default = 0}

\item{contact_matrix_set}{Contact matrices used in simulation. Default =
NULL, which will generate this based on the \code{country}.}

\item{R0}{Basic Reproduction Number. Default = 3}

\item{tt_R0}{Change time points for R0. Default = 0}

\item{beta_set}{Alternative parameterisation via beta rather than R0.
Default = NULL, which causes beta to be estimated from R0}

\item{time_period}{Length of simulation. Default = 365}

\item{dt}{Time Step. Default = 0.1}

\item{day_return}{Logical, do we want to return outut after each day rather
than each dt. Default = FALSE}

\item{replicates}{Number of replicates}

\item{init}{Data.frame of initial conditions. Default = NULL}

\item{seed}{Random Number Seed.}

\item{prob_hosp}{probability of hospitalisation by age.
Default = c(0.001127564, 0.000960857, 0.001774408, 0.003628171,
0.008100662, 0.015590734, 0.024597885, 0.035377529,
0.04385549, 0.058495518, 0.08747709, 0.109730508,
0.153943118, 0.177242143, 0.221362219, 0.267628264)}

\item{prob_severe}{Probability of developing severe symptoms by age.
Default = c(3.73755e-05, 3.18497e-05, 5.88166e-05, 0.000120264,
0.000268514, 0.000516788, 0.00081535, 0.001242525,
0.001729275, 0.002880196, 0.00598205, 0.010821894,
0.022736324, 0.035911156, 0.056362032, 0.081467057)}

\item{prob_non_severe_death_treatment}{Probability of death from non severe
treated infection.
Default = c(0.0125702, 0.0125702, 0.0125702, 0.0125702,
0.0125702, 0.0125702, 0.0125702, 0.013361147,
0.015104687, 0.019164124, 0.027477519, 0.041762108,
0.068531658, 0.105302319, 0.149305732, 0.20349534)}

\item{prob_non_severe_death_no_treatment}{Probability of death in non severe
hospital inections that aren't treated}

\item{prob_severe_death_treatment}{Probability of death from severe infection
that is treated. Default = rep(0.5, 16)}

\item{prob_severe_death_no_treatment}{Probability of death from severe infection
that is not treated. Default = rep(0.95, 16)}

\item{p_dist}{Preferentiality of age group receiving treatment relative to
other age groups when demand exceeds healthcare capacity.}

\item{dur_E}{Mean duration of incubation period (days). Default = 4.6}

\item{dur_IMild}{Mean duration of mild infection (days). Default = 2.1}

\item{dur_ICase}{Mean duration from symptom onset to hospitil admission (days).
Default = 4.5}

\item{dur_get_ox_survive}{Mean duration of oxygen given survive. Default = 9.5}

\item{tt_dur_get_ox_survive}{Times at which dur_get_ox_survive changes
(Default = 0 = doesn't change)}

\item{dur_get_ox_die}{Mean duration of oxygen given death. Default = 7.6}

\item{dur_not_get_ox_survive}{Mean duration without oxygen given survive.
Default = 4.75}

\item{dur_not_get_ox_die}{Mean duration without  oxygen given death.
Default = 3.8}

\item{dur_get_mv_survive}{Mean duration of ventilation given survive.
Default = 11.3}

\item{tt_dur_get_mv_survive}{Times at which dur_get_mv_survive changes
(Default = 0 = doesn't change)}

\item{dur_get_mv_die}{Mean duration of ventilation given death. Default = 10.1}

\item{dur_not_get_mv_survive}{Mean duration without ventilation given
survive. Default = 5.65}

\item{dur_not_get_mv_die}{Mean duration without ventilation given
death. Default = 1}

\item{dur_rec}{Duration of recovery after coming off ventilation. Default = 3.4}

\item{hosp_bed_capacity}{General bed capacity. Can be single number or vector if capacity time-varies.}

\item{ICU_bed_capacity}{ICU bed capacity. Can be single number or vector if capacity time-varies.}

\item{tt_hosp_beds}{Times at which hospital bed capacity changes (Default = 0 = doesn't change)}

\item{tt_ICU_beds}{Times at which ICU bed capacity changes (Default = 0 = doesn't change)}

\item{seeding_cases}{Initial number of cases seeding the epidemic}
}
\value{
Simulation output
}
\description{
Run the explicit SEEIR model
}
\details{
All durations are in days.
}
\examples{
\dontrun{
pop <- get_population("Afghanistan")
m1 <- run_explicit_SEEIR_model(R0 = 2,
population = pop$n, dt = 1,
contact_matrix_set=contact_matrices[[1]])
}
}
