% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc_utils.R
\name{sample_pmcmc}
\alias{sample_pmcmc}
\title{Sample PMCMC}
\usage{
sample_pmcmc(
  pmcmc_results,
  burnin = 0,
  n_chains,
  log_likelihood = calc_loglikelihood,
  n_trajectories = 10,
  n_particles = 100,
  forecast_days = 0
)
}
\arguments{
\item{pmcmc_results}{output of \code{run_mcmc_chain}; The results from the PMCMC run -- can have mutliple chains.}

\item{burnin}{integer; Number of iterations to discard from the start of MCMC run. Default = 0}

\item{n_chains}{number of chains that considered. Should inherent from pmcmc.}

\item{log_likelihood}{function to calculate log likelihood, must take named parameter vector as input,
allow passing of implicit arguments corresponding to the main function arguments.
Returns a named list, with entries:
  - $log_likelihood, a single numeric
  - $sample_state, a numeric vector corresponding to the state of a single particle, chosen at random,
  at the final time point for which we have data.
  If NULL, calculated using the function calc_loglikelihood.}

\item{n_trajectories}{interger; Number of trajectories to be returned. Integer. Default = 10.}

\item{n_particles}{integer; Number of particles to be considered in the particle filter. Default = 100}

\item{forecast_days}{integer; number of days being forecast. Default = 0}
}
\value{
\describe{
  \item{trajectories}{A 3-dimensional array of trajectories (time, state, tranjectories).}
  \item{sampled_PMCMC_Results}{The parameters chosen when sampling from the \code{pmcmc} posteriors}
  \item{inputs}{A list of model inputs.}
  }
}
\description{
Sample from the posterior probability results produced by \code{run_mcmc_chain}
to select parameter set. For each parmater set sampled, run particle
filter with \code{num_particles} and sample 1 trajectory
}
